deleteTable.php has been created.
The deleteTable.php page has been created with its respective CSS styles. In this, the user will be able to select a table from those he has created (visible on the page itself). When selecting one, the user will be prompted for confirmation to make sure they don't delete the table by mistake.