addRow page added. Small style changes to pages related to user tables.
An extra page called addRow has been created. In this page, when pressing the button on the page where the fields of the table are shown to the user, the form for introducing the fields of a new line of the table will be created. This implementation has been added to separate modifications of the user's table structure into individual pages.

At the same time, certain CSS style parameters of various pages have been modified and the footer has been added to the pages that lacked it.