Changes in the creation and display of user tables. First sketch of the deleteRow.php page.
The creation of user tables has been modified, as well as their display on the respective page. These tables now have a unique identifier, which will automatically increment as the user adds fields to the table of their choice. Note: The user will never have to specify a primary identifier.

At the same time, the first draft of the deleteRow.php page has been created, which will serve for the user to delete the rows he considers from the table he chooses.