Update deleteRow.php
The deleteRow.php page has been updated so that it allows deleting a row depending on its identification number that the user chooses. The user will be asked for confirmation when performing the action.

CSS styles have also been added according to the rest of the aesthetics of the application.