Update addTable.php, deleteRow.php, and addRow.php.
The pages shown in the title of the commit have been updated to favor a better integration of the elimination of rows by the user.

The use of IDs with the implementation of the MySQL AUTO_INCREMENT function has been discarded, given the incompatibilities with the auto update of table IDs after deleting an intermediate row from them.

The designation of a custom ID has been reintroduced in the section on creating the tables, which will later be used to be the field entered by the user to delete a specific row.

The CSS styles for various pages have been slightly modified.